server/requirements.txt
fastapi==0.95.1
google-api-python-client==2.86.0
redis==4.5.5
pandas==2.0.1
reportlab==4.0.0
openpyxl==3.1.2
uvicorn==0.22.0
pypdf2==3.0.1
matplotlib==3.7.1
python-dotenv==1.0.0
pydantic==1.10.8
httpx==0.24.1
